allow configuring the preferred solidity compiler
This change allows the configuration of the preferred Solidity compiler,
either lambda or local, by selecting the appropriate implementation
based on the value of the environment variable `SOLIDITY_COMPILER`.
This commit implements`SolcLambda` and `SolcLocal` classes to
represent the Lambda and local compiler implementations respectively.
The `ISolidityCompiler` interface is used to define the common methods
for both implementations.
