+++ services/server/src/server/controllers/verification/verification.common.ts
@@ -34,69 +34,22 @@ import { id as keccak256str } from "ethers";
 import { ForbiddenError } from "../../../common/errors/ForbiddenError";
 import { UnauthorizedError } from "../../../common/errors/UnauthorizedError";
 import { ISolidityCompiler } from "@ethereum-sourcify/lib-sourcify";
-import {
-  LambdaClient,
-  InvokeCommand,
-  InvokeCommandInput,
-} from "@aws-sdk/client-lambda";
+import { SolcLambda } from "../../services/compiler/lambda/SolcLambda";
+import { SolcLocal } from "../../services/compiler/local/SolcLocal";
 
-class Solc implements ISolidityCompiler {
-  private lambdaClient: LambdaClient;
-
-  constructor() {
-    if (
-      process.env.AWS_REGION === undefined ||
-      process.env.AWS_ACCESS_KEY_ID === undefined ||
-      process.env.AWS_SECRET_ACCESS_KEY === undefined
-    ) {
-      throw new Error("Solc compiler must be configured");
-    }
-    // Initialize Lambda client with environment variables for credentials
-    this.lambdaClient = new LambdaClient({
-      region: process.env.AWS_REGION,
-      credentials: {
-        accessKeyId: process.env.AWS_ACCESS_KEY_ID,
-        secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
-      },
-    });
+let selectedSolidityCompiler: ISolidityCompiler;
+switch (process.env.SOLIDITY_COMPILER) {
+  case "lambda": {
+    selectedSolidityCompiler = new SolcLambda();
+    break;
   }
-
-  public async compile(
-    version: string,
-    solcJsonInput: JsonInput,
-    forceEmscripten: boolean = false
-  ): Promise<CompilerOutput> {
-    const response = await this.invokeLambdaFunction(
-      JSON.stringify({ version, solcJsonInput, forceEmscripten })
-    );
-    return this.parseCompilerOutput(response);
-  }
-
-  private async invokeLambdaFunction(payload: string): Promise<any> {
-    const params: InvokeCommandInput = {
-      FunctionName: "compile",
-      Payload: payload,
-    };
-
-    const command = new InvokeCommand(params);
-    const response = await this.lambdaClient.send(command);
-
-    if (!response.Payload) {
-      throw new Error(
-        "Error: No response payload received from Lambda function"
-      );
-    }
-
-    return response;
-  }
-
-  private parseCompilerOutput(response: any): CompilerOutput {
-    const res = JSON.parse(Buffer.from(response.Payload).toString("utf8"));
-    return res.body as CompilerOutput;
+  case "local":
+  default: {
+    selectedSolidityCompiler = new SolcLocal();
   }
 }
 
-export const solc = new Solc();
+export const solc = selectedSolidityCompiler;
 
 export function createCheckedContract(
   metadata: Metadata,
